+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Episode list scroll lag
+- Sleep timer persists off-screen
 
 ## [v0.5.2] - 2026-07-05